+++ src/if/adpcm_tables.h
@@ -9,4 +9,4 @@ extern "C" uint32_t ADPCM_DitherState[2];
 extern "C" const uint16_t ADPCM_StepTable[89];
 
 /// @brief ADPCM step index table for 4-bit indices
-extern "C" const int16_t ADPCM_IndexTable_4bit[8];
+extern "C" const int8_t ADPCM_IndexTable_4bit[8];
